TRAC #29608 v0.0.1 - Nov 2014 - Jan 2015
This has been a long time in coming...

I first raised TRAC #29608 a year ago today, soon after WordPress 4.0 was released ( Benny, Sep 4, 2014 .
I started developing this solution with a plan to discuss it at WordCamp London Contributor day
in Nov 204.

I recently released an update to the new logic as part of oik-css v0.8.0
This plugin is dependent upon oik for some of its logic
and oik-css for the new filter functions.

It's purpose is to help you visualise how shortcodes are expanded using
the current WordPress code and the 'new' solution.